# ES2020 optimizations