--- a/src/execution/execute.ts
+++ b/src/execution/execute.ts
@@ -266,7 +266,7 @@ export function buildExecutionContext(
   let operation: OperationDefinitionNode | undefined;
   const fragments: ObjMap<FragmentDefinitionNode> = Object.create(null);
   for (let i = 0; i < document.definitions.length; i++) {
-    const definition = document.definitions[i]
+    const definition = document.definitions[i];
     switch (definition.kind) {
       case Kind.OPERATION_DEFINITION:
         if (operationName == null) {
@@ -432,9 +432,9 @@ function executeFields(
   const results = Object.create(null);
   let containsPromise = false;
 
-  const fieldsEntries = Array.from(fields.entries())
+  const fieldsEntries = Array.from(fields.entries());
   for (let i = 0; i < fieldsEntries.length; i++) {
-    const { 0: responseName, 1: fieldNodes } = fieldsEntries[i]
+    const { 0: responseName, 1: fieldNodes } = fieldsEntries[i];
     const fieldPath = addPath(path, responseName, parentType.name);
     const result = executeField(
       exeContext,
@@ -1230,7 +1230,9 @@ function executeSubscription(
     rootType,
     operation.selectionSet,
   );
-  const {0:responseName, 1:fieldNodes} = Array.from(rootFields.entries())[0];
+  const { 0: responseName, 1: fieldNodes } = Array.from(
+    rootFields.entries(),
+  )[0];
   const fieldName = fieldNodes[0].name.value;
   const fieldDef = schema.getField(rootType, fieldName);