--- a/src/execution/values.ts
+++ b/src/execution/values.ts
@@ -76,7 +76,8 @@ function coerceVariableValues(
   onError: (error: GraphQLError) => void,
 ): { [variable: string]: unknown } {
   const coercedValues: { [variable: string]: unknown } = {};
-  for (const varDefNode of varDefNodes) {
+  for (let i = 0; i < varDefNodes.length; i++) {
+    const varDefNode = varDefNodes[i]
     const varName = varDefNode.variable.name.value;
     const varType = typeFromAST(schema, varDefNode.type);
     if (!isInputType(varType)) {
@@ -161,7 +162,8 @@ export function getArgumentValues(
   const argumentNodes = node.arguments ?? [];
   const argNodeMap = keyMap(argumentNodes, (arg) => arg.name.value);
 
-  for (const argDef of def.args) {
+  for (let i = 0; i < def.args.length; i++) {
+    const argDef = def.args[i]
     const name = argDef.name;
     const argType = argDef.type;
     const argumentNode = argNodeMap[name];