--- a/src/jsutils/keyValMap.ts
+++ b/src/jsutils/keyValMap.ts
@@ -23,7 +23,8 @@ export function keyValMap<T, V>(
   valFn: (item: T) => V,
 ): ObjMap<V> {
   const result = Object.create(null);
-  for (const item of list) {
+  for (let i = 0; i < list.length; i++) {
+    const item = list[i]
     result[keyFn(item)] = valFn(item);
   }
   return result;