--- a/src/jsutils/toObjMap.ts
+++ b/src/jsutils/toObjMap.ts
@@ -13,9 +13,9 @@ export function toObjMap<T>(
   }
 
   const map = Object.create(null);
-  const entries = Object.entries(obj)
+  const entries = Object.entries(obj);
   for (let i = 0; i < entries.length; i++) {
-    const { 0: key, 1: value } = entries[i]
+    const { 0: key, 1: value } = entries[i];
     map[key] = value;
   }
   return map;