--- a/src/utilities/buildASTSchema.ts
+++ b/src/utilities/buildASTSchema.ts
@@ -49,7 +49,8 @@ export function buildASTSchema(
   const config = extendSchemaImpl(emptySchemaConfig, documentAST, options);
 
   if (config.astNode == null) {
-    for (const type of config.types) {
+    for (let i = 0; i < config.types.length; i++) {
+      const type = config.types[i]
       switch (type.name) {
         // Note: While this could make early assertions to get the correctly
         // typed values below, that would throw immediately while type system