--- a/src/utilities/buildClientSchema.ts
+++ b/src/utilities/buildClientSchema.ts
@@ -85,9 +85,9 @@ export function buildClientSchema(
   );
 
   // Include standard types only if they are used.
-  const stdTypes = [...specifiedScalarTypes, ...introspectionTypes]
+  const stdTypes = [...specifiedScalarTypes, ...introspectionTypes];
   for (let i = 0; i < stdTypes.length; i++) {
-    const stdType = stdTypes[i]
+    const stdType = stdTypes[i];
     if (typeMap[stdType.name]) {
       typeMap[stdType.name] = stdType;
     }