@IvanGoncharov 2022-08-11
@JoviDeCroock Thanks for PR and research.
Can you please rebase, since I added a few related improvements on `main`?
I also made an experiment to see, if we can keep modern syntax without a performance drop.
I published it as #3691, what do you think about it?

BTW, your branch is still faster in some tests but slower in others:
<img width="587" alt="image" src="https://user-images.githubusercontent.com/8336157/184211193-7d808055-69b6-475b-9f7a-ed15f7b92cee.png">
Maybe rebasing will fix this.