@JoviDeCroock 2022-08-11
@IvanGoncharov #3691 optimizes some problematic ones, you might want to add a plugin like https://babeljs.io/docs/en/babel-plugin-transform-destructuring or the one from react-hooks because they transform `const [x, y] = useState` to `const { 0: x, 1: y }` there as well. I think that might be the last performance gap we have between our benches, A last point of optimization might be avoiding using `extends` on the `Map` built-in.

If you want I can also close this as your transform approach looks better to me, thank you!